fsyjq-wechat/requirements.txt
astroid==1.6.1
autopep8==1.3.4
certifi==2018.1.18
chardet==3.0.4
decorator==4.2.1
Django==1.11
django-cors-headers==2.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
httpie==0.9.9
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.2.15
jedi==0.11.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.12
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycodestyle==2.3.1
Pygments==2.2.0
PyJWT==1.5.3
pylint==1.8.2
pytz==2017.3
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
wrapt==1.10.11
